Sacrifice specificity for sensitivity in branch pattern
The idea here is not to validate a branch's name, but to parse it from the
git status output. Git will not allow a branch to be created with an
invalid name. Thus we create a simple(ish) regex that will recognize any
valid branch, while not worrying about each branch naming rule at
http://git-scm.com/docs/git-check-ref-format#_description.