UX: Don't include --no-columns flag on `git branch`.
This flag was implemented in Git 1.7.11.  Older versions of Git are still in
use - for example, Ubuntu LTE 12.04 is still using 1.7.9.x.  Previously in
those environments, anything relying on `git branch` would fail with an
unsupported flag error.

Since it is likely uncommon to explicitly enable column output for Git globally,
and since the default is --no-column in version >1.7.11, we'll leave this
flag up to the user to include in their global_flags in GitSavvy.sublime-settings.

Closes #158.