Fix: Branch dashboard failed to show branch when commit includes "..".
The regular expression for matching the branch name included a greedy
negative look-ahead that consumed the entire line.  It was included
as a strictly-correct attempt to rule out branch names that included
two consecutive period characters.

This was an unnecessary change, however, because the condition is
impossible to reach (if we trust the git command-line tools).  When
reading a list of branch names, there will never be a branch with two
consecutive periods, because that branch could never have been created.

Resolves #204.