Enhancement: Fork UX in rebase-conflict scenario when one side is deleted.
Merge-tool launching is disallowed when one side has been deleted - a message
will be displayed to the user.  When choosing "use version from your commit"
or "use version from new base" in the rebase dashboard, and when that option
is that the file has been deleted, deleted the file in the working directory
and index instead of checking the file out (which would error).

Closes #267.