Feature: show and edit branch descriptions
Adds a new feature to show branch descriptions (obtained from local git config,
see https://git-scm.com/docs/git-branch for details) in the branches dashboard.

This is useful when you can't have meaningful branch names and only have them
named based only on the issue number in your issue tracker (for example,
"feature/AB-1234") - when you have a bunch of them.

Also adds a way to easily edit the descriptions from the dashboard, binded to
the E key with a short description in the documentation.