Fix: Pushing a single tag would not work in tag dashboard.
Closes #319.

Often when using the UI templates, there is no new-line at the end of a given
region (for example, `local_tags` in the tags dashboard).  For that reason,
the valid region is bounded by the end of the line, not including the new-line
character.

However, we were previously using `view.full_line` which _does_
include the new-line character.  Therefore, the region to test would end one
character (the new line) beyond the valid region.