Fix: display the "git binary not found" error message only once
This avoids blocking the user from using sublime by spamming her/him
with endless messages about git executable not found. It displays the alert
dialog once and continues to log further failures as exceptions.

This allows GitSavvy to start functioning once git is on path (or the
settings are updated to point to it's explicit location).

Fixes #551