Feature: rename GsLogBase -> LogMixin and enhance it
Firstly this renames GsLogBase command into LogMixin, and replaces
the inherited classes with ``object`` only. This avoids confusion
when using the mixin and requires to explicitly specify parent classes
to inherit from.

This commit also allows LogMixin to function as both TextCommand and
WindowCommand mixin and adds support for specifying branch when running.