Enhancement: Cache settings object for multi get/set operations
The number of settings need to be applied while creating a view has grown over the time. Therefore it might be a good idea to use a local reference of the object to apply or read a batch of view settings rather than calling `view.settings()` method several times.

In some cases the syntax is assigned before settings applied and sometimes after. This commit changes the code in order to always apply settings first.