Fix theme generator for sublime-color-schemes
Fixes #1029

All sublime JSON files can contain comments and the last item of a
dictionary or list is allowed to have a trailing comma. This makes
them invalid JSON files, which causes python's json library to raise
parse errors.

Hence this commit replaces the
 - json.loads with sublime.decode_value()
 - json.dumps with sublime.encode_value()

and removes the obsolete json import.