Add symlink support on Windows
The Issue
=========

Git for Windows does not enable support for symlinks by default as it
requires administrative privileges and causes a couple of edge cases.

Hence calling `git rev-parse --show-toplevel` is not able to resolve
junctions or symlinks and does not return the real repository path, if
the working directory is accessed via junction/symlink.

Another issue is caused by python itself, as its `os.path.realpath()`
function does not resolve junctions/symlinks on Windows as well.

The Motivation
==============

When working on `sublimehq/Packages` the only way to test changes of
single syntaxes is to create a junction of a Package (e.g. Python) in
ST's `Data/Packages` path. Otherwise syntax tests fail and changes of
the syntax-definition or any other module are not tracked by ST.

There may be other use cases for working with symlinks.

The Fix
=======

This commit enables GitSavvy to support this kind of workflow by
borrowing the `realpath patch` and the algorithm to find the repo's
toplevel path from GitGutter.

The fix consists of two parts:

1. A patched `realpath()` function, which is applied if Windows
   supports the underlying `_getfinalpathname()` function.

   On MacOS/Linux the default `os.path.realpath()` is called as it
   already supports symlinks.

2. A local alternative for the `git rev-parse --show-toplevel` command
   which resolves the repository path without calling any git command.

As a result any git command is now called with the real repository
path even if the working directory is accessed via a junction/symlink on all OS's.