Make behavior of common.util.realpath()
As noted the `common.util.realpath()` is borrowed from GitGutter which uses
the empty return value to detect the requested path does not exist or symlinks
could not be resolved.

This commit therefore changes the behavior of that function to match the
originial `os.path.realpath()` behavior in such edge cases.

If '' or `None` is provided as path -> current directory is returned
If `path` can not be resolved -> `path` is returned as is.
If `path` is resolved, the real path is returned.