Fix failing dashboard test case
Some further investigation revealed that `os.path.realpath()` translates
`/not/here/` to `C:\not\here` on Windows. When I did my last tests about how
that function works with none-existing paths I used a not existing windows-like
path. Hence the function seemed to return the input as is. The test cases use
unix like paths instead.

This commit changes the patch to fallback to `os.path.realpath()` if
FileNotFoundError is raised by `_getfinalpathname()` so that the patch should
now behave kindly like the original function in all edge case scenarios.