Go to next hunk after refresh if the view was empty before
This is a fairly good heuristic. If the diff was empty the cursor is
positioned at the EOF. Going to the next hunk will thus jump to the first
one.

Before, we just hardcoded to jump after init but since diff views can be
reused and become empty anytime this is just smarter.