Optimize diff view startup
GsDiffCommand either creates a new diff view, or it finds a suitable one
already open and then just focuses it. In both cases we want to make sure
that the view *eventually* shows the most recent diff.

The previous implementation had the following problems:

- `git diff` did run on the main thread (#1092)
- Reusing a diff view called `git diff` twice

We now generally rely on the `on_activated_async` event to update the view
lazy. The init command either creates or focuses a view but it never calls
for a 'refresh' imperatively.

Note that most of the time we actually want a blocking, sync update (to keep
the view and the repo state in sync) so we add a switch to the refresh command.

Fixes #1092