Fix branch panel logic if `ignore_current_branch`
The observable side effect of the bug is that `git branch`, called via
`get_current_branch_name()`, is asked twice e.g. if you want to checkout a
branch.

Usually we fill in `selected_branch` with the current branch if not set, but
we of course don't need to do that if `ignore_current_branch` is True.

Also simplified `get_pre_selected_branch_index`.

Fixes #1109