Rewrite status bar updater
Rewrite throttler code for the status bar updates.

The old throttler had a 100ms wait time but this is not a functionality we
need here. E.g. if you create a new view, Sublime puts two events on the queue,
namely 'on_new' and 'on_activated'. The new thottler allows only one task
per view on the worker thread to be executed. The most recent one wins.

Fixes #1094