Save repo_path on the GitCommand instance for status checks
Fixes #1093

We add a setter to `self.repo_path`, and use it in the statusbar updater code
to reduce `git --show-toplevel` calls for unsaved buffers.  This is the most
conservative way of adding another cache for now because it must be set
manually and it is only stored per instance.