Factor out getting a repo path for a view
We introduce a new global store `repo_paths` which maps directories to
repositories they belong to and factor out a useful function
`repo_path_for_view`.

We do not use the view settings as a cache, we still read the key
`git_savvy.repo_path` which is important for our special views.

Overall this greatly reduces `git --show-toplevel` commands at the risk
of a stale cache *if* the user creates sub-repos or deletes a repo. Here the
user would have to restart Sublime to clear the cache whereby before it was
always enough to close and reopen a view.