Fix reloader code
The old reloader did actually not reload the window and text commands which is
very annoying. We could fix it ourself, but there is no real minimal fix for it,
or just use the code from @randy3k's `AutomaticPackageReloader`.

The following is a copy of
https://github.com/randy3k/AutomaticPackageReloader/blob/
b568fcdf4e208d403f0fe5ee967016d589efa765/reloader/reloader.py
under MIT but @randy3k is also a member of GitSavvy.

Added the entry point function `reload_plugin`.