Subscribe status dashboard to the state
By doing that, we save one 'status' call per activate, and we get background
updates for free. E.g. if you have the status dashboard open in another tab,
it will automatically update in the background prior activating it.

Notable change: We call the 'close' handler 'pre_close' so that the view is
still attached to a window. (We want to `unsubscribe` before being detached.)
We still remove the interface instance 'post'-close.