Follow the path on the main thread
Since following the path queries the view *a lot* it is way faster to do this
on the main thread because it blocks the view. If we see very large paths
regulary and with it freezes, we must implement a way to shift this calculation
onto the worker thread if it takes to long.