Try to focus a file when status changes
In the dashboard, it is easy that the cursor gets into meaningless
land after an action. Likely, if after a commit the workspace gets
clean, or after it was clean, it becomes dirty again.

Before, we only moved the cursor `on_new_dashboard`. Now, we determine
after each render if the cursor is still on a meaningful line and if
not move it to the next file.

Couple of notes:

- For the inital draw, this is faster, because we focus
  immediately after the render which is just way earlier than
  `on_new_dashboard`.

- We only jump to files, excluding stash entries, which would become
  annoying. (We don't want to jump to the stashes section after commit
  for example.)

- I slightly modified the syntax to make this change work. Because
  I don't want to introduce visual regressions I use weird scope names
  like `meta.git-savvy.status.stash.name`. These are all `meta` and no
  theme will have them defined. I also don't know the naming rules
  here in Sublime land. However, generally the syntax should help
  selecting the right things, and I expect we use such selectors
  in the actions.