Simplify `view_is_transient`
In Sublime, there is always a "hidden" view i.e. when you just start up
Sublime, or e.g. close all views. This "hidden" view does not have a
tab initially, but it gets one as soon as you start typing.

We want to draw the basic branch information in the status bar in these
cases just as normal. For widgets, we never want to draw the status
bar.

We used `window.get_view_index` to detect widgets before but this is
hard to read and comprehend. We thus switch to the apparent check
`view.settings().get('is_widget')` to check if a given window is in fact
a widget.