Do not listen for the "load" event for the status bar
The "load" event typically fires with an "activated" event later.
I.e. when opening a file to edit. In this typical case, we receive two
events which is a waste. (The current throttler only filters >2 events
so it didn't catch here for us.)

The "load" event otherwise fires for preview views for which we do
not want to update the status bar anyway. We filter these events in
`view_is_transient`.

We hence remove the event handler because it has no purpose anymore.