Refactor diff parsing into multiple types
The base type is `TextRange = namedtuple('TextRange', 'text a b')`
which makes everything mildly natural.  Note that we parse the diff
from a string now, so we basically swap `view.find_all` with
`re.finditer`.

In the tests, we have to change the expected value from `(30, 34)`
to `(41, 46)`. The lefthand shifts by 11 chars because we have to add
`+++ b/fooz\n` to our view mock. The righthand shifts by 12(!) chars
because we now include the trailing '\n' in the returned Region.