Redo commit panel logic
How we implemented the `draw_info_panel` logic was upside down because
we thought we were in complete control (via the [m] key) of showing
and hiding that panel. But we're not since the user can issue hide and
show commands via other key bindings or the panel menu.

The new implementation is thereby driven by intercepting these commands
and acting upon them (by implementing `on_window_command`). Our own
toggle on the other side just fires such hide and show panel commands.
This is relatively straight forward, and easier to reason about.

How and when the panel opens and closes automatically can be fine-tuned
after this refactoring. Two things changed already though,

1. We now handle the `incremental_find` panel special. T.i. we take it
as a intermediate panel and upon closing it (via [ESC] or [enter]) we
immediately show the panel again (if it was open before).

2. When the user had a typical long-lived panel open while editing, e.g.
a SublimeLinter error panel or Terminus session, we now open the commit
panel when activating the graph view, and switch *back* to the previously
open panel when the user switches the views.

This commit also introduces some fine-tuning about what we run on the
UI thread and what on the worker. Generally speaking, we want some blocking
UI behavior to show the user some feedback immediately and e.g. to reduce
multiple redraws in short bursts of time. We thus start the graph command
(usually creating the view, setting its title etc.) on the main thread
now. We also open the panel immediately, synchronous (if needed) to
have the normal layout trashing which happens when opening a panel in
the same go as creating the new view.

The result is that we fastly show a new blank view splitted with the
panel, and then get one next redraw actually filling both the view and
the panel at the same time. Note that this change also affects the
log panel which relies on the same commit panel.

A small fix in `GsBranchesLogGraphCommand` was necessary because it
uses the graph mixin we change here. A proper factoring will come later.