Reload plugins when actual reloading failed
Fixes #1205

Before we reload we `unload_module` every plugin. In case of a hard
error while reloading (e.g. often a syntax error), we actually try
to bring the old modules back via `reload_missing` but did not register
the actual plugins. We thus have all the code somehow (inconsistent)
loaded but none of the commands for the plugins which actually raised.
(In GitSavvy we only have *one* plugin, so actually no commands are
registered after that.)

With this patch we call `reload_plugin` again after failures. The goal
is to keep at least the reload command itself working.

This is good for developers but could be bad for users. However, we
really don't expect reload failures for end-users.