Rollback all modules
Hard to imagine what we're doing here.

I try to rollback all old modules in place. This works as intended
during short testing. If I introduce two changes in two files, and one
of them has a syntax error. None of the changes are applied. (Before
we had a mixed state where the one good file had the changes, and the
one with the syntax error (and its imports) obviously was not reloaded.)