Reuse graph views
Fixes #1239

Here we actually go a bit different than for the other views in that
we don't store a in memory mapping from some id to view. If we do so
we would also have to garbage collect at some point, and in case of
restarts/reloads we need a fallback like this one here to refresh that
map.

So we just enumerate all possible views and look for one that seems
okay enough to reuse.