Reorder view setup and "refresh"
For the reuse case, this is significant because the panel opens on
`focus_view` which will trigger already a panel update. (After that
we navigate to the followed symbol which probably updates the panel
again.) So we have two panel redraws where we need one.

The code move for the create case has no runtime behavior change.