Trigger status refresh in main global_events
This is so that the status bar "controller" participates on the
IGNORE_NEXT_ACTIVATE signal.

We let go the first `on_activated` event even if we come from a widget,
for example the Command Palette.  This is because usually the user opens
a file using `ctrl+p`, and we must set our status in that case.

T.i. for normal views we trigger an status update on the first activate.
For our "special" views this is a no-op because they're still in a blank
state on the first *sync* activate event.  Neither any settings nor the
view syntax has been set at that point.