Rewrite throttler for the status bar
Fixes #1094

The old throttler had a 100ms wait time but this is not a functionality
we need here. E.g. if you create a new view, Sublime puts two events on
the queue, namely 'on_new' and 'on_activated'.

By using the `runtime.throttler` we ensure that only one task on the
worker wins and gets executed.

Note that the implementation does not differentiate between different
views, windows, or repo paths. So there is a risk to undercall now.
Although it is very unlikely to get multiple triggers from different
sources in exactly one task, we capture and merge such events until the
worker thread is ready again which can be at any time in the future.
Note though that the *last* event wins in such cases.