Fix: Check if view is a graph view before reusing it
We currently store (cache?) "repo_path" and "file_path" on each view
(if we used one of our getters like `self.repo_path` on them) which
makes the id not unique if 'all' is true and 'branches' None.

The fix is to explcitly check if the given view is a graph view
because only those are candidates for reuse anyway.