Optimize for simple refresh draws
If we just refresh the graph, i.e. `on_activated`, we assume that
most of the content stays the same. Since drawing is a bottleneck here
we slice the new graph from the end (where the old content is) and
only draw the changed content (usually at or near the top).