Stream content and draw while streaming
We optimize for first paint. Thus we will read the content streaming
in, process it on a separate thread and put actual diff tokens on a
queue. As soon as we get the first token we start the drawer on the
main thread which will start to apply the diff tokens to the view. As
soon as the first "follow" point has been drawn we switch to a slower
cooperative mode and yield back to the worker to draw the rest of
the diff.

Note that since `draw_` is itself a text command we will block both
the worker and the ui thread but only for short times.