Do not block UI thread endlessly for too big diffs
We blocked the UI thread until `did_navigate`. This is generally fast
and reduces visual flicker. It is also okay for small to medium sized
repos. (*I* get roughly 1sec paint time for 10.000 commits.)

For bigger repos, we don't want to stall or freeze Sublime too much.
Showing 50.000 commits is already not a very fast experience, and we
don't need to make it worse.