Proper abort for `draw_` function
Sublime doesn't call text commands for closed views. This means we
correctly aborted draining and drawing the queue but we missed killing
the potentially still running process.

We extract a common decorator `ensure_not_aborted` for this
functionality which we can use for `draw` and the inner `draw_`. Note
the order of the decorators on `draw_`.