Do not strip ANSI sequences from `self.git` output
This reverts ea98d366ff52d928d07211020694d1d199e70db7

The issue #1020 was about the output in the panels which was never
affected by this change. `self.git` returns the stdout usually for
our purposes. (Esp. the push output of said issue is not used by us.)

Stripping and normalizing the panel output is now done within the
panel functions.