Fix: Prevent infinte loop in the "trying" state
The intention here is (of course) to catch a very specific failure,
namely git not recognizing our new date format. Only then we change
the format and try to run the command again.

For every other error, we MUST raise as usual. Note that here we
construct a new `GitSavvyError` just for the flipped `show_panel` flag.
Just contructing this object already has the side-effect of showing
the panel.