Enhance error message if git command fails
For the streaming stdout we don't want to cache the result (obviously),
so we just store the state that we received something. Usually we
don't have when git errs and the desired effect of this commit is that
`stdout` becomes the empty string in that case and thus the error
message not showing "<SNIP>" when there was nothing snipped.