Read current graph directly from the view
First, storing large data on the view settings is a proper waste as it
gets persisted in the workspace.  Then we also have the problem for
aborted draws that the actual state of the view and the stored state
are not in sync and we just had to draw everything in these cases.

So we switch to just reading the *actual* content of the view.  This
happens to be just as fast as our storage model but without wasting
disk space. Note that we then have to `format_line` early t.i. before
diffing because the view content obviously is always formatted.