Flash status bar if no `JumpTo` positions can be found
Also: For non-empty selections just take the cursor at `begin()`. There
is no reason not to do that probably, and if you use a mouse you
sometimes select a char or so without further intention.

Note: Probably no need to consider `end()` as well because we filter
the results per filename anyway.