Respect view selection when following
We strictly "follow" t.i. set the cursor and scroll only for simple
selections. Selections are "simple" if there is only one cursor, and
maybe a selection over one line (e.g. a word).

Other selections are "complex", and we don't want to destroy such
constructions.

As a bonus we, when the cursor is somewhere on a line, we don't move
the cursor to the commit hash but keep the position. But if the cursor
is on the commit hash, we "normalize" and put it at the beginning of
the hash. This is useful if the column start of the hashes changes
because the graph characters before change. (Maybe you were on a nested
branch, which then after a commit becomes the leftmost branch.)