Block until viewport is painted
Before we painted as soon as we first jumped, e.g. to the symbol we
follow, t.i. on `just_navigated`. That's cheap but not really what we
want: We want to draw as soon as the viewport has been drawn completely.

We can only start measuring the "viewport" (here: `visible_region()`)
after we jumped because only then its extent will be reasonable. So we
now have to track `just_navigated` and `viewport_readied` (locally).

We clearly have a problem now when `did_navigate` never flips, and
with it `just_navigated` stays `False`. Then we run in blocking mode
til `TheEnd`. (E.g. the symbol we follow is gone.)

Aside: `visible_region` does not count empty space. T.i. if the view
is very short, less lines than fit, `visible_region` will only report
these exact lines. (This should be okay though for us since we do a
`>=` check.)