Only check for renames if necessary
`gs_show_file_at_commit` is usually invoked from a diff where we
extract the filename directly from the view or git output and thus know
it is valid for the given commit.

Only for the blame view it is actually necessary to follow renames.
We introduce a flag to model this behavior.

We extract the functionality out of `get_file_content_at_commit`
which is out of concern of the function anyway. There was also a bug
because we called said function with `self.file_path` instead of the
provided file_path the command was called with.