Ensure newline at end of graph
Generally a newline at the end of a file looks better. This one here
also ensures that the cursors stays left (for example at the empty
trailing newline) if a new drawn graph is shorter then the previous
one and the follow symbol has gone. This is more likely the case since
we allow dynamic filtering via `[f]`.