Do not offer "reset" on the head commit
Before this change we show e.g. "Reset 'dev' to 'dev'". The already
reads confusing. The standard action even presents a menu to select
the reset mode after that. This is then basically useless, for example
"--soft" is a noop, "--mixed" just cleans the stage.

The only useful option is "--hard", basically to clean the working dir.
But we don't provide any info (yet?) that we have a dirty status (gitk
inserts fake nodes for the unstaged and staged area) so removing it for
now seems reasonable.