Check commit hash against `newest_commit_for_file`
In the show commit view we either open the current file of the workspace
or a historical revision. Instead of just comparing the hash of the
shown commit with the HEAD, it makes more sense to ask git (basically
via `git log -1 -- file_path`) if the hunk we're looking at is the most
recent change to that file.