Revert `[o]` opens workspace file
For now revert the new behavior of `[o]` opening the current workdir
file if the change/hunk you're looking at is the most recent change
to that file.

We first just checked if the commit is the "HEAD" commit, then switched
to `newest_commit_for_file` which is more accurate behavior but shows
that this UX is confusing. (It is also relatively slow because we now
always need multiple git queries to answer the action.)

We state for now: either the user is in forensic mode, t.i. searching,
reading and looking around to understand a change; or the user sees a
change and then wants to edit (e.g. "fixup" that particular location).

The first interaction mode is bound to `[o]` again, for the latter we
have `[O]`. There is probably no need to mixmatch both.

The bevahior was first introduced in

  9c05c3d ("Open workspace file if shown commit is the HEAD commit")