Simplify `gs_custom`
Fixes #1212

There is technically no reason for a custom `Thread` class because
the default implements all features we need.

Rearrange the program so that running or not running in a separate
thread has the same features, t.i. setting a `status_message` or for
example `output_to_buffer` works in both cases now.

Especially, running `refresh_gitsavvy_interfaces` after the command
now works (#1212).