Run `gs_delete_branch` on the UI thread if branch is given
Deleting a branch is not very time-consuming, and generally this is
a response to an user action, so immediate is better than slightly
delayed.

Especially if we offer a forceful deletion (by showing a quick panel)
this makes a difference because the UI or cursor doesn't update in
between and thus not suggesting it will take long or it will or already
has succeeded.