Do not force a "cached" mode when reusing views
When reusing diff views it is often desired to not force a specific
"cached" mode but just switch to the view. For example, a user might
have a key binding to diff the current file. Here it is better to
just have *one* binding to open or focus the diff, and switch the
"cached" mode via `<TAB>`.

We thus make `in_cached_mode` a three-state toggle, t.i. not setting it
(`None`) becomes also meaningful in that it only focuses the view.

Fixes regression introduced by c74159fef38729bb2c179b74b6831991a31ce70f
("Update "cached" mode when reusing diff views"), thus related to #1299.