Simplify zero context diffs detection and handling
Zero context diffs need some fine tuning of the `b_start` value. We
can detect these here and thus don't need to pass that information via
an argument `zerodiff=True`. E.g. an `-U0` diff

```
@@ -145,0 +146,5 @@ def rewrite_hunks(hunks, zerodiff=True):
+def hunk_of_additions_only(hunk):
+    # type: (Hunk) -> bool
+    return hunk.a_length == 0 and hunk.b_length > 0
+
+
```

will become

```
@@ -145,2 +145,7 @@ def rewrite_hunks(hunks, zerodiff=True):

+def hunk_of_additions_only(hunk):
+    # type: (Hunk) -> bool
+    return hunk.a_length == 0 and hunk.b_length > 0
+
+

```

in the `-U1` case.  Note that `a_length=2` because of the two context
lines, also `b_start=145` instead of `146` showing that the `+-1`
tuning is not necessary if we have context lines.