Fix: Only select hunks if the cursor is within that hunk
A long standing bug is that the user can be on the line right above where
a hunk actually starts which then selected the immediately following hunk.
This is maybe acceptable if we then stage the whole hunk, but becomes a real
bug when only staging single lines. In that case `index_in_hunk=-1` so
it surprisingly stages the *last* line of the hunk.

This is a clasical off-by-one error we fix hereby.