Draw on the main thread
This is important for the cursor positioning because Sublime will
otherwise not draw the (moved) cursor correctly.

See #1203 for the same bug on a different view.

Removed the "viewport" fix which seems wrong. Especially *if* the user
actually wants to scroll to something "on the right", out of the max
viewport.