Fix: Abort on conflicted files
The removed `verify_not_conflict` was never called because parsing the
diff already threw an exception. It was always costly because a `status`
call is never cheap.

We remove `verify_not_conflict` and instead catch the parse error.

This similar to the code in `inline_stage_hunk` but does not yet share
enough code to reuse its `UnsupportedCombinedDiff`.