Optimize reusing inline diff views
This is in line with how we do this for the other views in that we don't
hold a mapping (here: `inline_diff_views`) but just iterate over the
views of the current window.

This has two advatanges: 1. We don't need to garbage collect if a view
closes. (But we didn't do this here anyway.) 2. We can reuse views after
a fresh start.