Revert closing the file if it has a clean diff
This actually likeable for example if you try to open the diff view
but there are actually no changes. But this naive implementation is
not good enough when it for example closes the view automatically if
you just staged the last remaining hunk.

We revert here; the actual feature could be implemented if we had
access to the general "status" information. E.g. prevent opening the
diff view *iff* the file is clean t.i. it neither has staged nor
unstaged changes.