Optimize stage toggling
When switching the stage mode we always made one additional call to git
to compute the correct next cursor. Here we refactor them away.

In cached mode: We actually need the non-cached diff to compute the next
line no. (So we basically made the same call twice.) Hence we read that
diff and pass it down to the "refresh" handler to avoid the second call.

In non cached mode: We actually used the reversed ("-R") diff of what
we already had in `diff_view_hunks`. So we just make a reverse version
of our algorithm instead.