Improve syntax file parsing
Fixes #1028

For a lot of files `yaml.safe_load` fails if the lib does not support
the used yaml format.  It is also very slow to parse all syntax files.

Since we only use relatively simple information here, t.i. a single
key, we just run a regex and extract the useful information.