Cache what the next commit is when going past in time
Generally `self.next_commit` is not memoizable because it makes a
`log commit..HEAD` call. But we can assume that if the user goes to
the previous commit and then back again (`[pnp]`), she wants to land
on the same commit again.

We cache on the view and thus never invalidate.