Fix: "rebase --continue" command
Here, our `GsRebaseContinueCommand` wants to be smart by commiting the
staged files, assumely so that the user doesn't need to edit the commit
message.

But this approach here leads to wrong commit messages because git added
comments in the prepared message, and `--no-edit` will not strip these.

Also, this will literally do a new commit, for example with a new date
and author et.al. which is not what we want.