After "pull" flash message if we were up to date
Note that the sibling command "push" sends all output to stderr which
we can't read using our `self.git` abstraction.  We omit it here for
simplicity.  (It is also not so important for "push" because a user
rather know if there are commits to push, but not necessarily knows if
there are new commits on the server.)