Add "fetch"/"push"/"pull" to the graph menu
Add "fetch", "pull", and "push" commands to the actions menu in the
log graph views.  Note that for "fetch" we try to read the remote
of the current branch, and fetch that remote, not only the branch.

Only add these items when on the HEAD of a branch.  (Here the menu
is usually rather short anyway since most options aren't applicable
here.)

Ideally, "push" would not be available if we're actually not behind,
t.i. up-to-date.