Implement smart copy for graph view
Sublime copies the whole line on "copy" if there is no selection.
For the graph view, we intercept this command and instead copy either
the current commit hash, the commit message, or a combination of both
to the clipboard ("<hash> (<message>)").

We cycle trhough these options.  If we copy something, we nicely flash
the saved portions of the view for visibility as well as the status bar.

If the user selected something or has multiple cursors, we just pass
through.