Implement "Line History View"
A "Line History" (aka "log wtf" or "log why") basically calls
"git log -L<some_file>,1:3".  Rarely used on the command line because of
its syntax, it is a super easy to write and then very useful tool in
the editor.  It is usually the faster (in terms of you get the info
you're looking for faster) "blame" and also the faster "File History"
because often you're only interested in some part of a file. Really
answering "Why is this line or section of code here?".

Comes with `[o]` to open the complete commit under the cursor;
`[O]` to open the file revision;
`[g]` to open the graph context.