Fix: Correctly draw gutter for initial revision of a file
`previous_commit` can return empty for the initial revision of a file.
`get_file_content_at_commit` though returns the current revision of a
file if the second arg is empty.

We special case here, and return the empty string as previous version
which in turns marks every line as newly added.