Marginally better movement after staging single lines
The code basically jumped to next "+" lines. This is only okay when
working with complete hunks.  When staging single lines we just do
nothing now, and thus stay at the same row.

Note that we check `name()` to determine which command or code path
we're actually running in.  That is obviously not OOP.  But it looks
like we haven't figured positioning the cursor here at all because
most options now default to "do nothing" (`None`).