Switch from "line" to "row" mode when staging
In 92f7513 (Fix: Only select hunks if the cursor is within that hunk)
we actually fixed an off-by-one error in the wrong direction (for what
our data holds). Since `section_start` already stores a `Row`, and the
used `rowcol` also returns a `Row` it is correct to switch the algorithm
to just using rows.

Changing the algorithm means switching `<=` to `<` and vice-versa (since
it is off by one).  This also better shows that `section_end` is
exclusive.

E.g. for a one line change on the first line, we have `section_start=0`
and `section_end=1`.  Being on row 1, the user is already *past* the
change.  (It mimics how `row in range(start, end)` would behave.)

Slightly optimized how we read the current `sel()` and flash a message
if we can't handle the current selection.