Fix `gs_prev_hunk`
Jumping "back" was maybe too tricky implemented.  There is another bug
when we have two modifications at the end of a hunk.  Because we always
choosed the line start of `b` (t.i. `view.line(b).a` in the code) there
was a chance landed in another modification.

Keep it simple, and just use a different algorithm for jumping "prev".
Generally, we want to jump to the start of a hunk even when jumping
"up".  We now jump twice ging up but then just take the `a` part which
is now the first modification of the hunk we want.

Also more understandable, prevent all viewport moves directly in
`jump`.  We then either `restore()` or show the region we found anyway.