Fix: Take the `b` filename as a goto symbol
We switch here, and mark `b` instead of `a` as the filename for the
Goto funtionality.  This is (more) correct here because we generally
have the current workstate on the b-side.

Imagine you check in a new file or you renamed a file.  In the first
case, `a` doesn't point to any file, in the latter `a` points to the
old file.