Fix `row_offset` if the view has some folded content
`row_offset` tells you for example that a given row is the 20th
in the viewport.  The algorithm didn't count for hidden rows which
happens if the user folds the code.

We switch to comparing the y values of two vectors.  `viewport_position`
return the top `y`, and `text_to_layout` returns the current `y` the
cursor is on.