Bind `up|down` arrow keys by default for navigation
We always had a flag `arrow_keys_navigation` (Default: `False`!) here,
but for the graph view we make it the default but also don't show the
jump target "at center" reducing the scrolling.  The effect is that the
arrow keys pretty much behave like before, just skipping some context
graph lines and always preserving some context lines at the top or
bottom of the view.

Overall technically a breaking change but to the good.

Note that the class doc mentions the `compare_commit_view`. This is
incorrect since c39691a (Obsolete and remove special "commit_compare"
view).