Implement long wide jumps in the graph view
Before, we only jumped "wide" if on an edge commit.  On a series of
consecutive calls both navigate commands behaved the same.

We switch here to two strategies.  Either the immediate `next_dot` is
a "wide" jump, then we take it.  Otherwise, we jump "wide" to the last
commit in the series of commits.