Rewrite settings abstraction for performance and reliability
Fixes #1201
Fixes #1341
Based on but rewrites rejected #1209

Generally, we want to support project wide overrides for the settings.
We could access these settings using `view.settings()`. However, for
`WindowCommand`s we don't necessarily have an `active_view()` for
example if the "view" is actually a `Sheet` showing an image or (in
ST4) a HTML page.  (Note that `sublime.active_window()` always yields
but `window.active_view()` can return `None`.)

For simplicity, we switch to depend solely on a `Window` and read the
project settings directly using `window.project_data()`.  We implement
caches for both the project settings and the global app settings. (The
caches make a difference if we run commands *not* on the main thread
because we avoid the synchronization.)